went through game, need to add alot of stuff, commented out the bad stuff. Just need to add game front end logic and we'll be good. After that confirm the backend game logic, and the auth middleware
the game needs a rehaul over how it considers operations and undo, also requires an introduction to selected and updated value cards. the cards themselves need to be updated in the operation process. we need to store operations in an array. and these operations will just reflect the game state. the key is when operations are updated - only when the selected items are a card, an operation, and another card. then we save, and all undo does is go back that operation.

OR it can dselect a selected number, or operation. otherwise it undoes from the operation stack, and the selection is on the first number. maybe theres another way but yeah we just track state when operations are fulfilled, then wait for operations til they're number and operator and number. and we have to actually change the values in the cards and track that in the state, and they're exact card location and

so whether an operation is selected is the flag of whether or not we have to make an append in operations state, cause if its a number its a new operation.

And the backend isn't going to track operations. That's just too much. we'll track submissions and if the submissions are too fast they'll get flagged and have to be verified. which we should add a verification modal, but we can include it in the AuthProvider. Plus verification if definitely an official version release feature.